Refactor LocateRunThread to background_tasks module
